ai_assistent/requirements.txt
openai>=1.0.0
redis>=5.0.0
aioredis>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0